+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 4.6.1 (November 2, 2023)
+
+### BUG FIXES
+
+*  Correct flaw in data source `okta_group` where name query matches multiple groups but did not consider exact match [#1799](https://github.com/okta/terraform-provider-okta/pull/1799). Thanks, [@monde](https://github.com/monde)!
+*  For resource `okta_idp_saml` set `status`, `sso_binding`, `sso_destination`, and `sso_url` during read context for proper import [#1558](https://github.com/okta/terraform-provider-okta/pull/1558). Thanks, [@monde](https://github.com/monde)!
+
 ## 4.6.0 (November 1, 2023)
 
 ### IMPROVEMENTS
@@ -15,7 +22,6 @@
 * Implement proper error for incorrect compound import input [#1785](https://github.com/okta/terraform-provider-okta/pull/1785). Thanks, [@monde](https://github.com/monde)!
 * Fix a panic in resource okta_resource_set [#1786](https://github.com/okta/terraform-provider-okta/pull/1786). Thanks, [@monde](https://github.com/monde)!
 * Correct change detection on resources okta_app_oauth_post_logout_redirect_uri and okta_app_oauth_redirect_uri [#1793](https://github.com/okta/terraform-provider-okta/pull/1793). Thanks, [@monde](https://github.com/monde)!
-
 
 ## 4.5.0 (October 17, 2023)
 
@@ -33,7 +39,6 @@
 
 ### IMPROVEMENTS
 * Add track all users argument to okta_group_memberships import [#1766](https://github.com/okta/terraform-provider-okta/pull/1766). Thanks, [@arvindkrishnakumar-okta](https://github.com/arvindkrishnakumar-okta)!
-
 
 ## 4.4.3 (October 09, 2023)